simple_dash_app/{{cookiecutter.project_name}}/requirements.txt
pandas<2.0
numpy
dash
dash-bootstrap-components
dash-labs
glob2==0.7
SQLAlchemy==1.4.31
xlrd==2.0.1
alembic==1.8.1
openpyxl==3.0.10
XlsxWriter==3.0.3
cx_Freeze
waitress
idna
cx_Logging
uvicorn
gunicorn
PyMySQL
urllib3
polars
python-dotenv
cookiecutter